Had trouble getting some things to compile, so I removed some legacy classes that have been replaced. Everyting compiles now, but some tests fail after updating everything to use new Matrix and Vector classes.

+++ tests/cpp_unit/BeamTools/BTTrackerTest.cc
@@ -143,7 +143,7 @@ TEST_F(BTTrackerTest, SolenoidTest_u) {
         BTTracker::integrate(u, &x[0], _sol, BTTracker::u, 0.1, 1.);
       }
     }
-    catch(Squeal squee) {EXPECT_EQ(p, 3);}  // p of particle 3 is 0 => squeal
+    catch(Squeal squee) {EXPECT_EQ(p, (size_t) 3);}  // p of particle 3 is 0 => squeal
     for(size_t i=0; i<8; ++i) {
       ASSERT_NEAR(results_sol_u[p*8+i], x[i], 1e-3) << "particle " << p 
           << " var " << _var[i] << "\nParticle coords: " << Print8Vector(&x[0]);
@@ -155,7 +155,7 @@ TEST_F(BTTrackerTest, SolenoidTest_u) {
                                       origin, CLHEP::HepRotationY(180*degree));
       }
     }
-    catch(Squeal squee) {EXPECT_EQ(p, 3);}  // p of particle 3 is 0 => squeal
+    catch(Squeal squee) {EXPECT_EQ(p, (size_t) 3);}  // p of particle 3 is 0 => squeal
     for(size_t i=0; i<8; ++i) {
       ASSERT_NEAR(_particles[p][i], x[i], 1e-3) << "particle " << p 
           << " var " << _var[i] << "\nParticle coords: " << Print8Vector(&x[0]);
